Move Hydrogen timeline view model to other view model
